requirements.txt
numpy~=2.4.0
matplotlib~=3.10.8
scipy~=1.16.3
tqdm~=4.67.1
#PyQt5~=5.15.11
pandas~=2.3.3
PyQt6~=6.10.1
PyQt6_sip~=13.10.3
python-docx~=1.2.0
pdf2docx
